+++ src/Tests/Kephas.Services.Tests/AmbientServicesExtensionsTest.cs
@@ -16,7 +16,7 @@ namespace Kephas.Tests
     using NUnit.Framework;
 
     /// <summary>
-    /// Test class for <see cref="AmbientServicesExtensions"/>.
+    /// Test class for <see cref="AppServiceCollectionExtensions"/>.
     /// </summary>
     [TestFixture]
     [SuppressMessage("StyleCop.CSharp.DocumentationRules", "SA1600:ElementsMustBeDocumented", Justification = "Reviewed. Suppression is OK here.")]