+++ src/Tests/Kephas.Services.Tests/AutofacAmbientServicesTest.cs
@@ -4,7 +4,7 @@ using System.Diagnostics.CodeAnalysis;
 using NUnit.Framework;
 
 /// <summary>
-/// Test class for <see cref="AmbientServices"/>.
+/// Test class for <see cref="AppServiceCollection"/>.
 /// </summary>
 [TestFixture]
 [SuppressMessage("StyleCop.CSharp.DocumentationRules", "SA1600:ElementsMustBeDocumented", Justification = "Reviewed. Suppression is OK here.")]