+++ src/Kephas.Services/release-notes.md
@@ -41,6 +41,8 @@ NEW code
 var logManager = AppServiceCollection.TryGetServiceInstance<ILogManager>();
 ```
 
+* Breaking change: renamed `AmbientServices` to `AppServiceCollection`.
+
 ## Export factory
 * Breaking change: Removed `ExportFactory(Func<Tuple<TContract, Action>> factory)` constructor.
 * Breaking change: Removed `IExportFactory.CreateExport`. Replace with `IExportFactory.CreateExportedValue`.